+++ pkg/tuic/tuic.go
@@ -3,6 +3,7 @@ package tuic
 import (
 	"errors"
 	"net/url"
+	"strings"
 
 	"github.com/aceberg/unbox/internal/check"
 	"github.com/aceberg/unbox/pkg/tls"
@@ -20,16 +21,22 @@ func Parse(raw string) (*TUIC, error) {
 		return nil, err
 	}
 
-	q := u.Query()
+	id := strings.TrimSpace(u.User.Username())
+	err = check.ValidateUUID(id)
+	if err != nil {
+		return nil, err
+	}
 
 	res := &TUIC{
 		Type:     "tuic",
 		Tag:      u.Fragment,
 		Server:   u.Hostname(),
 		Port:     portInt,
-		UUID:     u.User.Username(),
+		UUID:     id,
 		Password: u.User.Username(),
 	}
+
+	q := u.Query()
 
 	t, ok := tls.Get(q)
 	if ok {